Material Export: refactor common export code into base class
