+++ unity-assetpackage/Assets/UsdUnitySdk/Scripts/IO/Materials/ShaderExporterBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2fba685d627db44c8943ce945b666ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
